homogenize multi task id queries, fix jcesr
