+++ emmet-api/emmet/api/routes/molecules/molecules/query_operators.py
@@ -171,35 +171,6 @@ class DeprecationQuery(QueryOperator):
         return {"criteria": crit}
 
 
-class MultiTaskIDQuery(QueryOperator):
-    """
-    Method to generate a query for different task_ids
-    """
-
-    def query(
-        self,
-        task_ids: str | None = Query(
-            None, description="Comma-separated list of task_ids to query on"
-        ),
-    ) -> STORE_PARAMS:
-        crit = {}
-
-        if task_ids:
-            # TODO: do task_ids need to be converted to ints?
-            crit.update(
-                {
-                    "task_ids": {
-                        "$in": [task_id.strip() for task_id in task_ids.split(",")]
-                    }
-                }
-            )
-
-        return {"criteria": crit}
-
-    def ensure_indexes(self):  # pragma: no cover
-        return [("task_ids", False)]
-
-
 class MultiMPculeIDQuery(QueryOperator):
     """
     Method to generate a query for different root-level mpcule_id values
